+++ src/components/modules/components/combiners/combinator/styles.scss
@@ -1,0 +1,4 @@
+.combinator {
+    display: flex;
+    flex-flow: row wrap;
+}
