+++ src/components/modules/components/generators/oscillator/components/OscillatorControllers/index.ts
@@ -1,0 +1,1 @@
+export { default } from './OscillatorControllers';
