+++ src/components/modules/components/generators/oscillator/components/OscillatorWaveSelection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './OscillatorWaveSelection';
