--- src/components/modules/components/generators/waveTableOscillator/index.ts
@@ -1,1 +1,0 @@
-export { default } from './WaveTableOscillatorComponent';
